this commit pulled out the office-bldg and office-group into their own components
