+++ client/components/app/app.component.jsx
@@ -5,8 +5,8 @@ import { Stats, Stars, Sky, HTML, MapControls, StandardEffects } from 'drei';
 
 import FlipButton from '../flip-button/flip-button.component.jsx';
 import Ground from '../ground/ground.component.jsx';
-import FrontEndCity from '../front-end/front-end-city.component.jsx';
-import Login from '../Login.jsx';
+import City from '../city/city.component.jsx';
+import Login from '../login/login.component.jsx';
 
 import './app.style.css';
